database не получаеться!!!

+++ app/src/main/java/com/example/dagger2newexample/cache/model/RemoteKeysDao.kt
@@ -12,7 +12,7 @@ interface RemoteKeysDao {
     suspend fun insertAll(remoteKeys: List<RemoteKeys>)
 
     @Query("SELECT * FROM remote_keys WHERE repoId = :repoId")
-    suspend fun remoteKeysRepoId(repoId: Long): RemoteKeys?
+    suspend fun remoteKeysRepoId(repoId: Int): RemoteKeys?
 
     @Query("DELETE FROM remote_keys")
     suspend fun clearRemoteKeys()
